aprinter-teensy3.cpp: Use float instead of double.

+++ aprinter/printer/aprinter-teensy3.cpp
@@ -150,7 +150,7 @@ using PrinterParams = PrinterMainParams<
     24, // LookaheadBufferSize
     10, // LookaheadCommitCount
     ForceTimeout, // ForceTimeout
-    double, // FpType
+    float, // FpType
     Mk20ClockInterruptTimer_Ftm0_Ch0, // EventChannelTimer
     Mk20Watchdog,
     Mk20WatchdogParams<2000, 0>,
